+++ src/main/java/com/ssssogong/issuemanager/domain/Comment.java
@@ -3,9 +3,13 @@ package com.ssssogong.issuemanager.domain;
 import com.ssssogong.issuemanager.domain.account.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Getter
