+++ src/main/resources/application.yml
@@ -11,7 +11,7 @@ spring:
     driver-class-name: org.h2.Driver
   jpa:
     hibernate:
-      ddl-auto: update
+      ddl-auto: create
     properties:
       hibernate:
   #        show_sql: true
